fix: moved scalibr version to /version/ edit: improve error propagation
