+++ version/version.go
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package scalibr
+// Package version contains the current osv-scalibr version
+package version
 
 // ScannerVersion is the current version of the scanner that's displayed in the scan results.
 const ScannerVersion = "0.3.1"
